Add sha256 hash function for when adding blocks. Should probably parse out the hex data from the object.to_string though when passing into hash function. Also added genesis block in blockchain constructor.

+++ src/blockchain/block.cr
@@ -2,7 +2,7 @@ require "./transaction"
 
 class Block
     property index : Int32
-    property timestamp : Float64
+    property timestamp : Int64
     property transactions = [] of Transaction
     property proof : Int64
     property previousHash : String
